refactor data fetch to Components, useEffect
